refactor: extract uds connection common
