+++ rust/crates/core/src/common/mod.rs
@@ -1,0 +1,5 @@
+mod error;
+mod uds_channel;
+
+pub use error::UdsChannelError;
+pub use uds_channel::UdsChannel;
