+++ rust/crates/core/src/lib.rs
@@ -1,3 +1,4 @@
+mod common;
 pub mod db;
 mod embeddings;
 pub mod stt;
